fix review pending tab, showing reviews, cleaned up redundant code
